+++ Demo/YYTextDemo/NSString+YYAdd.m
@@ -74,10 +74,10 @@
 }
 
 - (BOOL)containsEmoji {
-    return [self containsEmojiForSystemVersion:[UIDevice currentDevice].systemVersion.floatValue];
+    return [self containsEmojiForSystemVersion:[UIDevice currentDevice].systemVersion.doubleValue];
 }
 
-- (BOOL)containsEmojiForSystemVersion:(float)systemVersion {
+- (BOOL)containsEmojiForSystemVersion:(double)systemVersion {
     // If detected, it MUST contains emoji; otherwise it MAY not contains emoji.
     static NSMutableCharacterSet *minSet8_3, *minSetOld;
     // If detected, it may contains emoji; otherwise it MUST NOT contains emoji.
